-Added Manager Tab and Refactored Code
